Add files via upload
Added a display for level (increases by 1 every 10 rows cleared), and shows the number of rows cleared. Added logic to allow for the speed of bricks to go faster, and implemented the ability to pause. Also added a new key bind to move the bricks down faster and to drop the brick immediately with space.